logica do contexo de layouyr movida
