+++ src/main.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import listRoutes from './routers'
-import Layout from './layout'
+import LayoutPage from './layout'
 import './index.css'
 
 
@@ -11,7 +11,7 @@ ReactDOM.createRoot(document.getElementById('root')).render(
     <RouterProvider router={
       createBrowserRouter([
         { path:"/",
-          element: <Layout />,
+          element: <LayoutPage />,
           children: listRoutes
         }
         ])
